backend/requirements.txt
# WCE Campus Assistant - Backend Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.3
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0

# PDF Processing
PyPDF2==3.0.1
pymupdf==1.23.8

# Vector Database
chromadb==0.4.22

# Groq (LLM)
groq==0.4.2

# Embeddings (HuggingFace)
sentence-transformers==2.3.1
torch>=2.0.0

# Optional: Cohere Reranker
cohere==4.45

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Utilities
numpy>=1.24.0
tqdm==4.66.1
